fix: 로고 애니메이션 hook 생성 및 일부 코드 수정
